refactoring

+++ src/main/java/com/example/CinemaRoom/repository/SeatsRepository.java
@@ -4,5 +4,5 @@ import com.example.CinemaRoom.model.Seat;
 import com.example.CinemaRoom.model.Seats;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeatsRepository extends JpaRepository<Seats, Long> {
+public interface SeatsRepository extends JpaRepository<Seats, Integer> {
 }
